+++ src/main/java/basemod/patches/com/megacrit/cardcrawl/helpers/CardLibrary/AddSwitch.java
@@ -1,16 +1,24 @@
 package basemod.patches.com.megacrit.cardcrawl.helpers.CardLibrary;
 
 import basemod.BaseMod;
-import com.evacipated.cardcrawl.modthespire.lib.SpireInsertPatch;
-import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
+import com.evacipated.cardcrawl.modthespire.lib.*;
 import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.cards.AbstractCard.CardColor;
+import com.megacrit.cardcrawl.helpers.CardLibrary;
 import com.megacrit.cardcrawl.unlock.UnlockTracker;
+import javassist.CtBehavior;
 
-@SpirePatch(cls = "com.megacrit.cardcrawl.helpers.CardLibrary", method = "add")
-public class AddSwitch {
-	@SpireInsertPatch(rloc = 37)
-	public static void Insert(AbstractCard card) {
+@SpirePatch(
+		clz=CardLibrary.class,
+		method="add"
+)
+public class AddSwitch
+{
+	@SpireInsertPatch(
+			locator=Locator.class
+	)
+	public static void Insert(AbstractCard card)
+	{
 		CardColor color = card.color;
 		switch (color) {
 			case RED:
@@ -24,6 +32,17 @@ public class AddSwitch {
 				if (UnlockTracker.isCardSeen(card.cardID)) {
 					BaseMod.incrementSeenCardCount(color);
 				}
+		}
+	}
+
+	private static class Locator extends SpireInsertLocator
+	{
+		@Override
+		public int[] Locate(CtBehavior ctMethodToPatch) throws Exception
+		{
+			Matcher finalMatcher = new Matcher.MethodCallMatcher(UnlockTracker.class, "isCardSeen");
+			int[] lines = LineFinder.findAllInOrder(ctMethodToPatch, finalMatcher);
+			return new int[]{lines[lines.length - 1]};
 		}
 	}
 }